Added auth to the tasks endpoints

+++ app/routers/tasks.py
@@ -1,48 +1,64 @@
-from fastapi import HTTPException,APIRouter,Query
+from fastapi import HTTPException,APIRouter,Query,Depends
 from datetime import datetime,date
 from ..database import SessionDep
+from ..auth import get_current_user
+from ..utils import get_current_user_object
 from ..models.sql_models import TasksBase,TasksUpdate,TasksDB
 from sqlmodel import select,desc,or_
 
 router = APIRouter()
 
 @router.post("/tasks/create")
-def create_task(task:TasksBase,session:SessionDep):
+def create_task(task:TasksBase,session:SessionDep, user:str=Depends(get_current_user)):
+    user_db = get_current_user_object(session,user)
+    if not user_db:
+        raise HTTPException(status_code=403,detail="User is not valid")
     task_db =TasksDB(**task.model_dump())
+    task_db.user = user_db
     session.add(task_db)
     session.commit()
     session.refresh(task_db)    
     return task_db
 
 @router.get('/tasks/{id}')
-def get_task(id:int, session:SessionDep):
-    
+def get_task(id:int, session:SessionDep,user:str=Depends(get_current_user)):
+    user_db = get_current_user_object(session,user)
+    if not user_db:
+        raise HTTPException(status_code=403,detail="User is not valid")
     task = session.get(TasksDB,id)
     if not task:
         raise HTTPException(status_code=404, detail="The task is not avaiable")    
+    if task.user_id!=user_db.id:
+        raise HTTPException(status_code=404, detail="THis task does not belong to you")
     return task
 
 
-@router.get('/tasks/{date}')
-def get_by_date(date: date, session: SessionDep):
-    # Here, query the database to filter tasks by the specified date
-    start_of_day = datetime.combine(date, datetime.min.time())
-    end_of_day = datetime.combine(date, datetime.max.time())
-    tasks = session.exec(select(TasksDB).where(
-        TasksDB.created_at>=start_of_day,
-        TasksDB.created_at<= end_of_day
-        )).all()
-    return tasks
+# @router.get('/tasks/{date}')
+# def get_by_date(date: date, session: SessionDep):
+#     # Here, query the database to filter tasks by the specified date
+#     start_of_day = datetime.combine(date, datetime.min.time())
+#     end_of_day = datetime.combine(date, datetime.max.time())
+#     tasks = session.exec(select(TasksDB).where(
+#         TasksDB.created_at>=start_of_day,
+#         TasksDB.created_at<= end_of_day
+#         )).all()
+#     return tasks
 
 @router.get('/tasks')
 def get_tasks(session:SessionDep,
+              user:str=Depends(get_current_user),
               priority:bool=Query(default=False), 
               completed:bool=Query(default=None),
               keyword:str=Query(default=None),
               date: date=Query(default=None),
               offset:int=Query(default=0),
               limit:int=Query(default=100)):
+    
+    user_db = get_current_user_object(session,user)
+    if not user_db:
+        raise HTTPException(status_code=403, detail="User is not valid")
     query = select(TasksDB)
+    query = query.where(TasksDB.user_id==user_db.id)
     
     if date:
         start_of_day = datetime.combine(date, datetime.min.time())
@@ -61,10 +77,17 @@ def get_tasks(session:SessionDep,
     
 
 @router.patch("/tasks/{id}")
-def update_task(id:int, u_task:TasksUpdate,session:SessionDep):
+def update_task(id:int, u_task:TasksUpdate,session:SessionDep,user:str=Depends(get_current_user)):
+    user_db = get_current_user_object(session,user)
+    if not user_db:
+        raise HTTPException(status_code=403, detail="User is not valid")
+
     task_db = session.get(TasksDB,id)    
     if not task_db:
         raise HTTPException(status_code=404, detail="There is no such post")
+    if task_db.user_id!=user_db.id:
+        raise HTTPException(status_code=404, detail="THis task does not belong to you")
+        
     task_data = u_task.model_dump(exclude_unset=True)
     for key,value in task_data.items():
         setattr(task_db,key,value)
@@ -74,10 +97,16 @@ def update_task(id:int, u_task:TasksUpdate,session:SessionDep):
     return task_db
 
 @router.delete('/tasks/{id}')
-def delete_task(id:int,session:SessionDep):
+
+def delete_task(id:int,session:SessionDep,user:str=Depends(get_current_user)):
+    user_db = get_current_user_object(session,user)
+    if not user_db:
+        raise HTTPException(status_code=403, detail="User is not valid")
     task = session.get(TasksDB,id)
     if not task:
         raise HTTPException(status_code=404, detail="There is no such task")
+    if task.user_id!=user_db.id:
+        raise HTTPException(status_code=404, detail="THis task does not belong to you")
     session.delete(task)
     session.commit()
     return {"data":"Task is deleted"}
